refactor(bin): minify needless structure of crate
