--- src/model/mod.rs
@@ -1,1 +1,0 @@
-pub mod env;
